MSI clock works on startup, but you can't use the ADC!
Now we can completely ditch the system_xxx files.  the generic startup will take care of .data
to ram and zeroing bss, and has a weak overrideable SystemInit() function.  You can use that
to set up clocks and flash and power supply modes if the system defaults are not appropriate