Properly support variable skip lengths for tx
This allows proper, understandable behaviour when dealing with digi's xbee series one modules. In their
default config, even in 802.15.4 mode, they insert a two byte "Maxstream" header between the 802.15.4
header and the data payload.  Allow specifying whether to try and fake this by inserting 2 extra bytes
when transmitting.  This is not used on RX, as it could depend on the node you are receiving from.